The mobile menu links are connected to their respective sections. 2. Hide the menu when clicking the button.
