+++ index.html
@@ -45,6 +45,7 @@
           </ul>
           <div class="vet-header__toggle">
             <button
+              id="burger-menu-btn"
               class="menu"
               onclick="showMenu()" 
               aria-label="Main Menu"
@@ -65,13 +66,13 @@
         <div class="vet-headder-mobile-menu-wrapper" id="mobileMenu"> 
           <ul>
               <li>
-                Testimonios
+                <a href="#testimonios">Testimonios</a>
               </li>
               <li>
-                Servicios
+                <a href="#servicios">Servicios</a>
               </li>
               <li>
-                Sobre mi
+                <a href="#sobreMi">Sobre mi</a>
               </li>
           </ul>
         </div> 
@@ -465,7 +466,7 @@
 
 
     <!-- final Testimonios -->
-    <script src="public/data//comment.js"></script>
-    <script src="main.js"></script>
+    
+    <script type="module" src="main.js"></script>
   </body>
 </html>
